requirements.txt
Pillow>=9.0.0
PyQt6>=6.4.0
pyperclip>=1.8.0